requirements.txt
aiogram==3.15.0
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.2
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
groq==1.2.0
